gcp/projects/cuttlefish-hackathon/hackathon-controller/ui/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
requests==2.31.0
google-auth>=2.23.3
